feat(catalogo): add DTO and improve CategoryService
 - Created DTO and updated types - Added empty constructor to Category - Updated pom.xml with TestContainers - CategoryRepository now supports existsByName - CategoryService with custom exceptions and clean response for frontend - Added initial validations in CategoryServiceTest

+++ comercio-catalogo/src/main/java/com/comercio/comercio_catalogo/category/CategoryService.java
@@ -2,30 +2,55 @@ package com.comercio.comercio_catalogo.category;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+
+import com.comercio.comercio_catalogo.category.dto.mapper.CategoryDTOMapper;
+import com.comercio.comercio_catalogo.category.dto.request.CategoryRequestDTO;
+import com.comercio.comercio_catalogo.category.dto.response.CategoryResponseDTO;
 
 @Service
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final CategoryDTOMapper categoryDTOMapper;
 
-    public CategoryService(CategoryRepository categoryRepository) {
+    public CategoryService(CategoryRepository categoryRepository, CategoryDTOMapper categoryDTOMapper) {
         this.categoryRepository = categoryRepository;
+        this.categoryDTOMapper = categoryDTOMapper;
     }
 
-    public List<Category> getAll() {
-        return categoryRepository.findAll();
+    public List<CategoryResponseDTO> getAll() {
+        List<Category> categories = categoryRepository.findAll();
+        if (categories.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Nenhuma categoria encontrada.");
+        }
+
+        return categories.stream()
+                    .map(categoryDTOMapper::toResponse)
+                    .toList();
     }
 
-    public Category getById(String id) {
-        return categoryRepository.findById(id).orElse(null);
+    public CategoryResponseDTO getById(String id) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,"Category not found with id: " + id));
+        return categoryDTOMapper.toResponse(category);
     }
 
-    public Category create(Category category) {
-        return categoryRepository.save(category);
+    public CategoryResponseDTO create(CategoryRequestDTO categoryDTO) {
+        if (categoryRepository.existsByName(categoryDTO.name())) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Category with name '" + categoryDTO.name() + "' already exists.");
+        }
+        Category category = categoryDTOMapper.toRequest(categoryDTO);
+        Category saved = categoryRepository.save(category);
+        return categoryDTOMapper.toResponse(saved);
     }
 
     public void delete(String id) {
+        if (!categoryRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Category not found with id: " + id);
+        }
         categoryRepository.deleteById(id);
     }
 }
